requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.2
numpy>=1.24.0
tqdm>=4.65.0
streamlit>=1.22.0